Move notifications to shares_base
