+++ backend/dataall/modules/s3_datasets_shares/tasks/dataset_subscription_task.py
@@ -11,7 +11,7 @@ from dataall.core.environment.services.environment_service import EnvironmentSer
 from dataall.base.db import get_engine
 from dataall.modules.shares_base.db.share_object_models import ShareObjectItem
 from dataall.modules.s3_datasets_shares.db.share_object_repositories import ShareObjectRepository
-from dataall.modules.s3_datasets_shares.services.share_notification_service import ShareNotificationService
+from dataall.modules.shares_base.services.share_notification_service import ShareNotificationService
 from dataall.modules.s3_datasets.aws.sns_dataset_client import SnsDatasetClient
 from dataall.modules.s3_datasets.db.dataset_location_repositories import DatasetLocationRepository
 from dataall.modules.s3_datasets.db.dataset_table_repositories import DatasetTableRepository
